% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{add_clust_info}
\alias{add_clust_info}
\title{Add cluster information}
\usage{
add_clust_info(node_color, trajectories)
}
\arguments{
\item{node_color}{data frame mapping node to cluster and corresponding color}

\item{trajectories}{List of trajectories computed from shortest paths.}
}
\value{
Dataframe of trajectory elements (trajElmnts)
and cluster trajectory (traj_cluster).
}
\description{
Add cluster information
}
