% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids}
\alias{centroids}
\title{Data frame of centroids, each row represents the mean lab values of
all observations in each node of `f_graph`.
Output of the training workflow to establish the topology from fake data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 18 columns.
}
\usage{
centroids
}
\description{
Data frame of centroids, each row represents the mean lab values of
all observations in each node of `f_graph`.
Output of the training workflow to establish the topology from fake data.
}
\keyword{datasets}
