% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{color_clust}
\alias{color_clust}
\title{Get cluster colors for nodes.}
\usage{
color_clust(f_sim_map, my_clusters)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{my_clusters}{Community clusters from igraph.}
}
\value{
Data frame of nodes and corresponding colors
based on the cluster each node belongs.
}
\description{
Get cluster colors for nodes.
}
