% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{color_graph}
\alias{color_graph}
\title{Color nodes of graph given coloring method.}
\usage{
color_graph(
  f_sim_map,
  f_graph,
  f_time,
  my_colors,
  method = c("basic", "clust_shade", "clust_color")
)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{f_graph}{igraph object, out put from graph.adjacency}

\item{f_time}{Data frame of the original data but
with only two columns: ID and val (time)}

\item{my_colors}{Character vector of hex values specifying
color palette for enrichment.}

\item{method}{Character string specifying the coloring method.
Can be 'basic', 'clust_shade' or 'clust_color'.}
}
\value{
Modified graph with colors at nodes.
}
\description{
Color nodes of graph given coloring method.
}
