% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{color_vertex}
\alias{color_vertex}
\title{Adjust colors of nodes using enrichment function.}
\usage{
color_vertex(f_sim_map, f_time, my_colors)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{f_time}{Data frame of the original data but
with only two columns: ID and val (time)}

\item{my_colors}{Character vector of hex values specifying
color palette for enrichment.}
}
\value{
Data frame of vertex ids and colors.
}
\description{
Adjust colors of nodes using enrichment function.
}
