% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{commu_clus}
\alias{commu_clus}
\title{Detect community structure based on edge igraph::betweeness.}
\usage{
commu_clus(f_graph, directed = FALSE, bridges = TRUE, ...)
}
\arguments{
\item{f_graph}{igraph object, out put from graph.adjacency}

\item{directed}{Logical. whether to calculate directed edge betweenness
for directed graphs. Ignored for undirected graphs.}

\item{bridges}{Logical. whether to return a list the edge removals
which actually a component of the graph.}

\item{\dots}{Additional arguments to pass to igraph::edge.betweenness.community.}
}
\value{
Community clusters from igraph.
}
\description{
Detect community structure based on edge igraph::betweeness.
}
