% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{convert_id_to_node}
\alias{convert_id_to_node}
\title{Assign observations (ids) to nodes in network.}
\usage{
convert_id_to_node(f_sim_map, processed_data)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{processed_data}{Dataframe. Processed from original data.}
}
\value{
Dataframe of id, node and covid_id.
}
\description{
Assign observations (ids) to nodes in network.
}
