% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-funcs.R
\name{cosine_sim_func}
\alias{cosine_sim_func}
\title{Calculate the pairwise similarity between rows of numeric matrices}
\usage{
cosine_sim_func(a, b = NULL)
}
\arguments{
\item{a}{Input matrix.}

\item{b}{Input matrix. Optional.
If `b` is NULL, use `a`, i.e. similarity between the rows of `a`
would be computed.}
}
\value{
Matrix of the similarity between the rows of a and
the rows of b.
}
\description{
Calculate the pairwise similarity between rows of numeric matrices
}
\examples{
m <- matrix(1:8, ncol = 4) # a 2x4 matrix
n <- matrix(9:16, ncol = 4) # a 2x4 matrix
cosine_sim_func(m)
cosine_sim_func(m, n) # a 4x4 similarity matrix
}
