% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_graph}
\alias{f_graph}
\title{igraph object (from `graph.adjacency`)
output of the training workflow to establish the topology from fake data.}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
f_graph
}
\description{
igraph object (from `graph.adjacency`)
output of the training workflow to establish the topology from fake data.
}
\keyword{datasets}
