% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_trajectories}
\alias{find_trajectories}
\title{Find trajectories in the MST
All the trajectory or choose the nodes.}
\usage{
find_trajectories(processed_data, f_sim_map, f_graph)
}
\arguments{
\item{processed_data}{Dataframe. Processed from original data.}

\item{f_sim_map}{TDAmapper object.}

\item{f_graph}{igraph object, output from graph.adjacency.}
}
\value{
List of trajectories computed from shortest paths.
}
\description{
Find trajectories in the MST
All the trajectory or choose the nodes.
}
\examples{
my_tda <- map_tda(scaled_lab_mat)
my_graph <- make_tda_graph(my_tda, processed_data, 'time')
my_trajs <- find_trajectories(processed_data, my_tda, my_graph)
head(my_trajs[[1]])
head(my_trajs[[2]])

}
