% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{get_graph_properties}
\alias{get_graph_properties}
\title{Get basic properties of graph}
\usage{
get_graph_properties(f_graph, f_sim_map = NULL, simplified = TRUE)
}
\arguments{
\item{f_graph}{igraph object, out put from graph.adjacency}

\item{f_sim_map}{Optional TDAmapper object}

\item{simplified}{Boolean. If TRUE, return a numeric vectors of
basic properties. If FALSE, return a more comprehensive list of
properties including maximal connected components and
number of elements in each node.}
}
\value{
List or numeric vector of different properties including
number of graph components, number of nodes, median degree,
edge density, etc.
}
\description{
Get basic properties of graph
}
\examples{
my_tda <- map_tda(scaled_lab_mat)
my_graph <- make_tda_graph(my_tda, processed_data, 'time')
get_graph_properties(my_graph)
}
