% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{make_graph}
\alias{make_graph}
\title{Make igraph object by calculating adjacency, resizing nodes,
weighting edges and color nodes.}
\usage{
make_graph(
  f_sim_map,
  f_time,
  color_method = "clust_color",
  my_colors = c("#00A3DD", "#60C659", "#FFBC21", "#FF7F1E", "#EF2B2D")
)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{f_time}{Dataframe of id and enrichment variable.
Default is time (performing time enrichment).
However, f_time can contain any other variable.}

\item{color_method}{Character string specifying the coloring method.
Can be 'basic', 'clust_shade' or 'clust_color'.}

\item{my_colors}{Character vector of hex values specifying
color palette for enrichment.}
}
\value{
igraph object of the graph output.
}
\description{
Make igraph object by calculating adjacency, resizing nodes,
weighting edges and color nodes.
}
