% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{make_tda_graph}
\alias{make_tda_graph}
\title{Make igraph object by calculating adjacency, resizing nodes,
weighting edges and color nodes.}
\usage{
make_tda_graph(
  f_sim_map,
  data,
  enrich_var,
  color_method = "clust_color",
  my_colors = c("#00A3DD", "#60C659", "#FFBC21", "#FF7F1E", "#EF2B2D")
)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{data}{Processe data frame of the original data.}

\item{enrich_var}{Character string of enrichment variable.
Any column name of the processed data frame.}

\item{color_method}{Character string specifying the coloring method.
Can be 'basic', 'clust_shade' or 'clust_color'.}

\item{my_colors}{Character vector of hex values specifying
color palette for enrichment.}
}
\value{
igraph object of the graph output.
}
\description{
Make igraph object by calculating adjacency, resizing nodes,
weighting edges and color nodes.
}
\examples{
my_tda <- map_tda(scaled_lab_mat)
make_tda_graph(my_tda, processed_data, 'time')

}
