% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{minspantree}
\alias{minspantree}
\title{Get minimum spanning tree from graph}
\usage{
minspantree(f_graph)
}
\arguments{
\item{f_graph}{igraph object, out put from graph.adjacency}
}
\value{
igraph object representing the minimumspanningtree
}
\description{
Get minimum spanning tree from graph
}
