% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{node_color}
\alias{node_color}
\title{Dataframe mapping node to cluster and corresponding color,
output of the training workflow to establish the topology from fake data.}
\format{
An object of class \code{data.frame} with 36 rows and 3 columns.
}
\usage{
node_color
}
\description{
Dataframe mapping node to cluster and corresponding color,
output of the training workflow to establish the topology from fake data.
}
\keyword{datasets}
