% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_trajectories}
\alias{out_trajectories}
\title{List of trajectories computed from shortest paths.
output of the training workflow to establish the topology from fake data.}
\format{
An object of class \code{list} of length 8.
}
\usage{
out_trajectories
}
\description{
List of trajectories computed from shortest paths.
output of the training workflow to establish the topology from fake data.
}
\keyword{datasets}
