% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processed_data}
\alias{processed_data}
\title{Processed data from sim_dat}
\format{
A data frame with 7207 observations and 19 variables:
\code{id}, \code{covid_id}, \code{time}, \code{AlanineAminoTransferase},
\code{AspartateAminotransferase}, \code{Albumin},
\code{CardiacTroponinHighSensitivity}, \code{CRP}, \code{Creatinine},
\code{Ddimer}, \code{Ferritin}, \code{Fibrinogen}, \code{LDH}, \code{Lymphocyte},
\code{Neutrophil}, \code{Procalcitonin}, \code{PT}, \code{Bilirubin},
\code{WBC}.
}
\usage{
processed_data
}
\description{
Processed data from sim_dat
}
\keyword{datasets}
