% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{resize_nodes}
\alias{resize_nodes}
\title{Resize the nodes of graph given the number of points it contains}
\usage{
resize_nodes(f_sim_map, f_graph, mult = 6, add = 8)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{f_graph}{igraph object, out put from graph.adjacency}

\item{mult}{Scalar as a multiplicative size rescale parameter.}

\item{add}{Scalar as an additive size rescale parameter.}
}
\value{
An igraph object of the modified f_graph.
}
\description{
Resize the nodes of graph given the number of points it contains
}
