% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scaled_lab_mat}
\alias{scaled_lab_mat}
\title{Matrix of the scaled lab values of \code{sim_dat}.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7207 rows and 16 columns.
}
\usage{
scaled_lab_mat
}
\description{
Matrix of the scaled lab values of \code{sim_dat}.
}
\keyword{datasets}
