% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-funcs.R
\name{shortest_paths_func}
\alias{shortest_paths_func}
\title{Find the shortest paths.}
\usage{
shortest_paths_func(x, mst_weights)
}
\arguments{
\item{x}{Pair of node ids from start node to end node.}

\item{mst_weights}{Output from igraph::mst.}
}
\value{
Vector of node ids of shortest paths.
}
\description{
Find the shortest paths.
}
