% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{Simulated data from github.com/aridag/TDA_PSEUDOTIME.}
\format{
A data frame with 200 observations and 21 variables:
\code{id}, \code{covid_id}, \code{day}, \code{AlanineAminoTransferase},
\code{AspartateAminotransferase}, \code{Albumin},
\code{CardiacTroponinHighSensitivity}, \code{CRP}, \code{Creatinine},
\code{Ddimer}, \code{Ferritin}, \code{Fibrinogen}, \code{LDH}, \code{Lymphocyte},
\code{Neutrophil}, \code{Procalcitonin}, \code{PT}, \code{Bilirubin},
\code{WBC}, \code{first_date}, and \code{time}.
}
\usage{
sim_dat
}
\description{
https://github.com/aridag/TDA_PSEUDOTIME/blob/master/MyDataSim.csv
}
\keyword{datasets}
