% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-funcs.R
\name{sim_intersection}
\alias{sim_intersection}
\title{Compute intersection similarity to assign each subject to
most similar trajectory}
\usage{
sim_intersection(x, y)
}
\arguments{
\item{x}{Numeric vector of node numbers
indicating an individual's trajectory}

\item{y}{Numeric vector of node numbers
indicating a general trajectory}
}
\value{
A scalar as measure of similarity
}
\description{
Compute intersection similarity to assign each subject to
most similar trajectory
}
\examples{
sim_intersection(c(1,2), c(1,3,4))
}
