% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-funcs.R
\name{sim_jaccard}
\alias{sim_jaccard}
\title{Compute Jaccard similarity to assign each subject to
most similar trajectory}
\usage{
sim_jaccard(x, y)
}
\arguments{
\item{x}{Numeric vector of node numbers
indicating an individual's trajectory}

\item{y}{Numeric vector of node numbers
indicating a general trajectory}
}
\value{
A scalar as measure of similarity
}
\description{
Compute Jaccard similarity to assign each subject to
most similar trajectory
}
\examples{
sim_jaccard(c(1,2), c(1,3,4))
}
