% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-funcs.R
\name{weight_edges}
\alias{weight_edges}
\title{Weight edges of a graph based on mean time of each edge.}
\usage{
weight_edges(f_sim_map, f_graph, f_time)
}
\arguments{
\item{f_sim_map}{TDAmapper object}

\item{f_graph}{igraph object, out put from graph.adjacency}

\item{f_time}{Data frame of the original data but
with only two columns: ID and val (time)}
}
\value{
An igraph object of the modified f_graph.
}
\description{
Weight edges of a graph based on mean time of each edge.
}
