% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{widen_i2b2}
\alias{widen_i2b2}
\title{Convert i2b2 observation dataframe from long to wide.}
\usage{
widen_i2b2(df)
}
\arguments{
\item{df}{i2b2 observation dataframe.}
}
\value{
Wide dataframe.
}
\description{
Convert i2b2 observation dataframe from long to wide.
}
\examples{
widen_i2b2(sim_dat)
}
