% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Compute similarity based on trajectory}
\usage{
compute_similarity(
  processed_data,
  node_color,
  trajectories,
  f_sim_map = NULL,
  verbose = TRUE
)
}
\arguments{
\item{processed_data}{Dataframe. Processed from original data.}

\item{node_color}{Dataframe mapping node to cluster and corresponding color.}

\item{trajectories}{List of trajectories computed from shortest paths.}

\item{f_sim_map}{TDAmapper object}

\item{verbose}{Boolean. Whether to print trajectory.}
}
\value{
List of two components:
dataframe with various similarity measures and
dataframe of the observations mapped to corresponding node and cluster.
}
\description{
Compute similarity based on trajectory
}
\examples{
my_tda <- map_tda(scaled_lab_mat)
my_graph <- make_tda_graph(my_tda, sim_dat, 'time')
my_trajs <- find_trajectories(sim_dat, my_tda, my_graph)
compute_similarity(sim_dat, my_graph$node_color, my_trajs, my_tda)

}
