% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Compute similarity based on trajectory}
\usage{
compute_similarity(
  processed_data,
  f_sim_map,
  f_graph,
  trajectories,
  verbose = TRUE
)
}
\arguments{
\item{processed_data}{Dataframe. Processed from original data.}

\item{f_sim_map}{TDAmapper object}

\item{f_graph}{igraph object, out put from graph.adjacency.}

\item{trajectories}{List of trajectories computed from shortest paths.}

\item{verbose}{Boolean. Whether to print trajectory.}
}
\value{
Dataframe with different similarity measures
}
\description{
Compute similarity based on trajectory
}
