% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-funcs.R
\name{cosine_sim_func}
\alias{cosine_sim_func}
\title{Calculate the pairwise similarity between rows of input matrix}
\usage{
cosine_sim_func(input_mat)
}
\arguments{
\item{input_mat}{Input matrix.}
}
\value{
Square matrix of the similarity. Diagonal values should be 1.
}
\description{
Calculate the pairwise similarity between rows of input matrix
}
\examples{
m <- matrix(1:8, ncol = 2) # a 4x2 matrix
cosine_sim_func(m) # a 4x4 similarity matrix
}
