% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_trajectories}
\alias{find_trajectories}
\title{Find trajectories in the MST
All the trajectory or choose the nodes.}
\usage{
find_trajectories(minspantreeweights, processed_data, f_sim_map, f_graph)
}
\arguments{
\item{minspantreeweights}{Output from igraph::mst.}

\item{processed_data}{Dataframe. Processed from original data.}

\item{f_sim_map}{TDAmapper object.}

\item{f_graph}{igraph object, out put from graph.adjacency.}
}
\value{
Dataframe with different similarity measures
}
\description{
Find trajectories in the MST
All the trajectory or choose the nodes.
}
