% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{Simulated data from github.com/aridag/TDA_PSEUDOTIME.}
\format{
A data frame with 23849 observations and 6 variables:
\code{siteid}, \code{patient_num}, \code{days_since_admission},
\code{concept_type}, \code{concept_code}, and \code{value}.
}
\usage{
sim_dat
}
\description{
https://github.com/aridag/TDA_PSEUDOTIME/blob/master/MyDataSim.csv
}
\keyword{datasets}
